chore: pull auth-related functions into separate module
