+++ src/neptune_fetcher/api/api_client.py
@@ -39,21 +39,9 @@ from typing import (
     Union,
 )
 
-from neptune_api import (
-    AuthenticatedClient,
-    Client,
-)
-from neptune_api.api.backend import (
-    get_client_config,
-    get_project,
-)
-from neptune_api.auth_helpers import exchange_api_key
+from neptune_api.api.backend import get_project
 from neptune_api.credentials import Credentials
-from neptune_api.models import (
-    ClientConfig,
-    Error,
-    ProjectDTO,
-)
+from neptune_api.models import ProjectDTO
 from neptune_retrieval_api.api.default import (
     get_multiple_float_series_values_proto,
     query_attribute_definitions_proto,
@@ -90,9 +78,12 @@ from neptune_fetcher.fields import (
     FloatPointValue,
 )
 from neptune_fetcher.util import NeptuneException
+from neptune_fetcher.utilities.net import (
+    create_auth_api_client,
+    get_config_and_token_urls,
+)
 
 API_TOKEN_ENV_NAME: Final[str] = "NEPTUNE_API_TOKEN"
-NEPTUNE_VERIFY_SSL: Final[bool] = os.environ.get("NEPTUNE_VERIFY_SSL", "1").lower() in {"1", "true"}
 
 # Disable httpx logging, httpx logs requests at INFO level
 logging.getLogger("httpx").setLevel(logging.WARN)
@@ -304,48 +295,6 @@ class _SeriesRequest:
     container_id: str
     include_inherited: bool
     after_step: Optional[float]
-
-
-@dataclass
-class TokenRefreshingURLs:
-    authorization_endpoint: str
-    token_endpoint: str
-
-    @classmethod
-    def from_dict(cls, data: dict) -> TokenRefreshingURLs:
-        return TokenRefreshingURLs(
-            authorization_endpoint=data["authorization_endpoint"], token_endpoint=data["token_endpoint"]
-        )
-
-
-def get_config_and_token_urls(
-    *, credentials: Credentials, proxies: Optional[Dict[str, str]]
-) -> tuple[ClientConfig, TokenRefreshingURLs]:
-    with Client(base_url=credentials.base_url, httpx_args={"mounts": proxies}, verify_ssl=NEPTUNE_VERIFY_SSL) as client:
-        config = get_client_config.sync(client=client)
-        if config is None or isinstance(config, Error):
-            raise RuntimeError(f"Failed to get client config: {config}")
-        response = client.get_httpx_client().get(config.security.open_id_discovery)
-        token_urls = TokenRefreshingURLs.from_dict(response.json())
-    return config, token_urls
-
-
-def create_auth_api_client(
-    *,
-    credentials: Credentials,
-    config: ClientConfig,
-    token_refreshing_urls: TokenRefreshingURLs,
-    proxies: Optional[Dict[str, str]],
-) -> AuthenticatedClient:
-    return AuthenticatedClient(
-        base_url=credentials.base_url,
-        credentials=credentials,
-        client_id=config.security.client_id,
-        token_refreshing_endpoint=token_refreshing_urls.token_endpoint,
-        api_key_exchange_callback=exchange_api_key,
-        verify_ssl=NEPTUNE_VERIFY_SSL,
-        httpx_args={"mounts": proxies, "http2": False},
-    )
 
 
 def backoff_retry(
